# docs/swagger.yaml
basePath: /api
definitions:
  controlerAuth.User:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  models.AddressGeo:
    description: AddressGeo represents the geocode result for an address.
    properties:
      suggestions:
        items:
          properties:
            data:
              properties:
                country:
                  type: string
                postal_code:
                  type: string
              type: object
            unrestricted_value:
              type: string
            value:
              type: string
          type: object
        type: array
    type: object
  models.AddressSearch:
    description: AddressSearch represents the search result for an address.
    properties:
      metro:
        items:
          properties:
            distance:
              type: number
            line:
              type: string
            name:
              type: string
          type: object
        type: array
      result:
        type: string
      source:
        type: string
    type: object
  models.GeocodeRequest:
    description: GeocodeRequest represents the request body for address geocoding.
    properties:
      lat:
        type: string
      lon:
        type: string
    type: object
  models.SearchRequest:
    description: SearchRequest represents the request body for address search
    properties:
      query:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API for address searching and geocoding using Dadata
    API.
  termsOfService: http://localhost:8080/swagger/index.html
  title: My API
  version: "1.1"
paths:
  /address/geocode:
    post:
      consumes:
      - application/json
      operationId: geocodeAddress
      parameters:
      - description: Geocode Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.GeocodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddressGeo'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Geocode an address
      tags:
      - geocode
  /address/search:
    post:
      consumes:
      - application/json
      operationId: searchAddress
      parameters:
      - description: Search Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SearchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddressSearch'
        "500":
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Search for an address
      tags:
      - search
  /login:
    post:
      consumes:
      - application/json
      operationId: loginUser
      parameters:
      - description: User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controlerAuth.User'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
        "401":
          description: Invalid credentials
      summary: Log in a user
      tags:
      - SingIn
  /register:
    post:
      consumes:
      - application/json
      operationId: registerUser
      parameters:
      - description: User
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controlerAuth.User'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Invalid request format
      summary: Register a new user
      tags:
      - SingUp
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"